"8:50am.
```
    let annotations expr (s: CharStream<_>) =
        let i = col s
        let inline expr_indent expr (s: CharStream<_>) = expr_indent i (<=) expr s
        pipe2 (expr_indent expr) (opt (expr_indent pp' >>. expr_indent expr))
            (fun a -> function
                | Some b -> binop TypeAnnot a b
                | None -> a) s
```

Let me start with this.

9:05am.

```
let inline expr_indent i op expr d = if op i (col d) then expr d else d.FailWith ExpectedIndentation

let annotations expr d =
    let i = col d
    let inline expr_indent expr d = expr_indent i (<=) expr d
    pipe2 (expr_indent expr) (opt (expr_indent colon >>. expr_indent expr))
        (fun a -> function
            | Some b -> binop TypeAnnot a b
            | None -> a) d
```

This should deal with it.

This actually typechecks now.

9:15am.

```
let process_parser_exprs exprs =
    let error_statement_in_last_pos = pfail StatementLastInBlock
    let rec process_parser_exprs on_succ = function
        | [ParserExpr a] -> on_succ a
        | [ParserStatement _] -> error_statement_in_last_pos
        | ParserStatement a :: xs -> process_parser_exprs (a >> on_succ) xs
        | ParserExpr a :: xs -> process_parser_exprs (l "" (unop ErrorNonUnit a) >> on_succ) xs
        | [] -> on_succ B

    process_parser_exprs preturn (List.concat exprs)
```

Typechecks as well. Some parts of the parser really had great work done on it.

9:25am.

```
let indentations statements expressions d =
    let i = col d
    let inline if_ op tr s = expr_indent i op tr s

    let inline many_semis expr = many (if_ (<) semicolon >>. if_ (<) expr)
    let inline many_indents expr = many1 (if_ (=) (pipe2 expr (many_semis expr) (fun a b -> a :: b)))
    many_indents ((statements |>> ParserStatement) <|> (exprpos expressions |>> ParserExpr)) >>= process_parser_exprs <| d
```

I am going to have the statements themselves return whether they are `ParserStatement` or `ParserExpr`.

The reason being is that this will allow me to fuse the `inl ... =` and `inl ... ->` into one. I'll also have to check the `inl` case in expressions, but it won't be such a difficulty there.

Let me leave this as such. Let me move on to the statements next."

+++ The Spiral Language/SpiralTypes.fs
@@ -42,7 +42,7 @@ type SpiralModule =
     }
 
 
-type PosKey = SpiralModule * int64 * int64
+type PosKey = SpiralModule * int * int
 
 type Pos<'a> = Position of PosKey * 'a with
     member x.Expression = match x with Position (_, expr) -> expr
